+++ src/main/java/chess/util/path/Web.java
@@ -1,0 +1,8 @@
+package chess.util.path;
+
+public class Web {
+
+    public static final String MAIN_PAGE = "/";
+    public static final String USER_HISTORY = "/user/name/:userName";
+    public static final String COMMAND_ACTION = "/game/command/:command";
+}
